refactor(model.state): 패키지명 변경
- state -> role
