+++ src/main/java/model/chessboard/ChessBoardFactory.java
@@ -8,14 +8,14 @@ import java.util.List;
 import java.util.Map;
 import model.piece.Color;
 import model.piece.PieceHolder;
-import model.piece.state.Bishop;
-import model.piece.state.King;
-import model.piece.state.Knight;
-import model.piece.state.Pawn;
-import model.piece.state.Queen;
-import model.piece.state.Role;
-import model.piece.state.Rook;
-import model.piece.state.Square;
+import model.piece.role.Bishop;
+import model.piece.role.King;
+import model.piece.role.Knight;
+import model.piece.role.Pawn;
+import model.piece.role.Queen;
+import model.piece.role.Role;
+import model.piece.role.Rook;
+import model.piece.role.Square;
 import model.position.Position;
 
 public class ChessBoardFactory {
