+++ src/main/java/model/piece/role/MultiShiftRole.java
@@ -1,4 +1,4 @@
-package model.piece.state;
+package model.piece.role;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,15 +8,15 @@ import model.piece.Color;
 import model.position.Position;
 import model.position.Route;
 
-public abstract class SingleShiftRole extends Role {
-    protected SingleShiftRole(Color color, ShiftPattern shiftPattern) {
+public abstract class MultiShiftRole extends Role {
+    protected MultiShiftRole(Color color, ShiftPattern shiftPattern) {
         super(color, shiftPattern);
     }
 
     @Override
     protected Route findRouteByDirection(Direction direction, Position source) {
         List<Position> sequentialPositions = new ArrayList<>();
-        if (source.isAvailablePosition(direction)) {
+        while (source.isAvailablePosition(direction)) {
             source = source.getNextPosition(direction);
             sequentialPositions.add(source);
         }
