+++ src/main/java/model/piece/role/Square.java
@@ -1,4 +1,4 @@
-package model.piece.state;
+package model.piece.role;
 
 import model.direction.Direction;
 import model.direction.ShiftPattern;
